move polygon coordinates to their own file
