+++ src/utils/polygon.hpp
@@ -2,14 +2,23 @@
 #define RGBDSLAM_UTILS_POLYGON_UTILS_HPP
 
 #include "../types.hpp"
-#include "coordinates/point_coordinates.hpp"
 #include <boost/geometry/geometry.hpp>
 #include <opencv2/core/mat.hpp>
 
 namespace rgbd_slam::utils {
 
-class WorldPolygon;
-class CameraPolygon;
+/**
+ * \brief Compute the projection of a point from the plane coordinate system to world
+ * \param[in] pointToProject The point to project to world, in plane coordinates
+ * \param[in] planeCenter The center point of the plane
+ * \param[in] xAxis The unit x vector of the plane, othogonal to the normal
+ * \param[in] yAxis The unit y vector of the plane, othogonal to the normal and u
+ * \return A 3D point corresponding to pointToProject, in world coordinate system
+ */
+vector3 get_point_from_plane_coordinates(const vector2& pointToProject,
+                                         const vector3& planeCenter,
+                                         const vector3& xAxis,
+                                         const vector3& yAxis);
 
 /**
  * \brief Describe a polygon by it's points in the polygon space
@@ -210,75 +219,9 @@ class Polygon
 };
 
 /**
- * \brief Describe a polygon in camera coordinates
+ * \return a polygon in screen space that represent the screen boundaries
  */
-class CameraPolygon : public Polygon
-{
-  public:
-    using Polygon::Polygon;
-    CameraPolygon(const Polygon& other) : Polygon(other) {};
-
-    /**
-     * \brief display the polygon in screen space on the given image
-     * \param[in] color Color to draw this polygon with
-     * \param[in, out] debugImage Image on which the polygon will be displayed
-     */
-    void display(const cv::Scalar& color, cv::Mat& debugImage) const noexcept;
-
-    /**
-     * \brief Project this polygon to the world space
-     * \param[in] cameraToWorld Matrix to go from camera to world space
-     * \return The polygon in world coordinates
-     */
-    [[nodiscard]] WorldPolygon to_world_space(const CameraToWorldMatrix& cameraToWorld) const;
-
-    /**
-     * \brief Compute the boundary points in screen coordinates
-     */
-    [[nodiscard]] std::vector<ScreenCoordinate> get_screen_points() const;
-
-    /**
-     * \brief Project this polygon to screen space
-     */
-    [[nodiscard]] polygon to_screen_space() const;
-
-    /**
-     * \brief Check that this polygon is visible in screen space
-     * \return true if the polygon is visible from the camera 1
-     */
-    [[nodiscard]] bool is_visible_in_screen_space() const;
-};
-
-/**
- * \brief Describe a polygon in world coordinates
- */
-class WorldPolygon : public Polygon
-{
-  public:
-    using Polygon::Polygon;
-    WorldPolygon(const Polygon& other) : Polygon(other) {};
-
-    /**
-     * \brief project this polygon to camera space
-     * \param[in] worldToCamera A matrix to convert from world to camera view
-     / \return This polygon in camera space
-     */
-    [[nodiscard]] CameraPolygon to_camera_space(const WorldToCameraMatrix& worldToCamera) const;
-
-    /**
-     * \brief Merge the other polygon into this one
-     * \param[in] other The other polygon to merge into this one
-     */
-    void merge(const WorldPolygon& other);
-
-    /**
-     * \brief display the polygon in screen space on the given image
-     * \param[in] worldToCamera A matrix tp convert from world to camera space
-     * \param[in] color Color to draw this polygon with
-     * \param[in, out] debugImage Image on which the polygon will be displayed
-     */
-    void display(const WorldToCameraMatrix& worldToCamera, const cv::Scalar& color, cv::Mat& debugImage) const noexcept;
-};
+Polygon::polygon get_static_screen_boundary_polygon() noexcept;
 
 } // namespace rgbd_slam::utils
